\name{download}
\alias{download}
\title{download the tables online}
\usage{
  download(directory)
}
\arguments{
  \item{directory}{the local directory for user to store
  data}
}
\value{
  no return and only download the tables
}
\description{
  Download all the census into local space.
}
\author{
  Yiming Li
}
\examples{
  download("D:/R/data")
}

