\name{help.census}
\alias{help.census}
\title{give the user the function list}
\usage{
  help.census()
}
\value{
  no return and only output the function list or data frame
}
\description{
  List all the functions and corresponding to description
  in the package.
}
\author{
  Yiming Li
}
\examples{
  help.census()
}

