\name{help.functions}
\alias{help.functions}
\docType{data}
\title{List of functions for helping}
\description{
This is a list of functions with the corresponding descriptions.
}
\usage{data(help.functions)}
\format{A data frame with 17 rows and 2 columns.}
\examples{
data(help.functions)
## maybe str(help.functions) ; plot(help.functions) ...
}
\keyword{datasets}
