\name{level.list}
\alias{level.list}
\docType{data}
\title{List of ordered hierarchy}
\description{
This is a list of hierarchies with the corresponding columns in the lookup table.
}
\usage{data(level.list)}
\format{A data frame with 6 rows and 2 columns.}
\examples{
data(level.list)
## maybe str(level.list) ; plot(level.list) ...
}
\keyword{datasets}
