\name{lookup1}
\alias{lookup1}
\docType{data}
\title{2011 OAs best-fitted to a number of output geographies}
\description{
2011 OAs best-fitted to a number of output geographies.
}
\usage{data(lookup1)}
\format{A data frame with 184170 rows and 11 columns.
}
\source{Office for National Statistics in UK.
}
\examples{
data(lookup1)
## maybe str(lookup1) ; plot(lookup1) ...
}
\keyword{datasets}
