\name{table.cells}
\alias{table.cells}
\title{get all the columns codes}
\usage{
  table.columns(table_code)
}
\arguments{
  \item{table_code}{the vector, special table codes}
}
\value{
  no return and only output them
}
\description{
  Output all the columns codes for special table code.
}
\author{
  Yiming Li
}
\examples{
  table.cells("KS101EW")
}

